requirements.txt
selenium
numpy
keras
tensorflow
pillow
SciPy